refactor: change guide gallery container size adapt
